requirements.txt
torch>=2.1
matplotlib
numpy
pillow
opencv-python
# choose ONE of the two lines below
jupyter-bbox-widget==0.6.0
# jupyter-bbox-widget @ git+https://github.com/roboflow/bbox-widget.git
sam-2 @ git+https://github.com/facebookresearch/segment-anything-2@main
